refactor: model → type 이름 수정
