+++ src/entities/shared-posts-filter/index.ts
@@ -1,2 +1,2 @@
-export * from './shared-posts-filter.model';
+export * from './shared-posts-filter.type';
 export * from './shared-posts-filter.hook';
