+++ src/entities/user/index.ts
@@ -1,1 +1,1 @@
-export type { User } from './user.model';
+export type { User } from './user.type';
