+++ src/entities/user/user.type.ts
