+++ src/features/auth/auth.atom.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil';
 
-import { type Auth } from './auth.model';
+import { type Auth } from './auth.type';
 
 export const authState = atom<Auth | null>({
   key: 'authState',
