+++ src/features/auth/auth.hook.ts
@@ -5,7 +5,7 @@ import { useRecoilState, useRecoilValue } from 'recoil';
 
 import { getUserData } from './auth.api';
 import { authState } from './auth.atom';
-import { type Auth } from './auth.model';
+import { type Auth } from './auth.type';
 import { authIsLogin } from './auth.selector';
 
 import { type User } from '@/entities/user';
