+++ src/features/shared/shared.dto.ts
@@ -1,4 +1,6 @@
-export interface GetSharedPostDTO {
+import { type SuccessBaseDTO } from '@/shared/types';
+
+export interface GetSharedPostsDTO extends SuccessBaseDTO {
   message: string;
   data: {
     content: PostContentObject[];
