+++ src/features/shared/shared.type.ts
