+++ src/shared/types/dto.ts
@@ -1,0 +1,11 @@
+export interface SuccessBaseDTO {
+  message: string;
+  data: object;
+}
+
+export interface FailureDTO {
+  status: number;
+  code: string;
+  errorMsg: string;
+  reason: string;
+}
