+++ src/shared/types/index.ts
@@ -1,2 +1,3 @@
+export * from './dto';
 export * from './rental-type';
 export * from './room-type';
